chore: moved numerical integration methods
