+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## [3.0.0]
 
  - **BREAKING**: Replaced the `Laguerre` type with `DurandKerner` (the latter is a more reliable root-finding algorithm for polynomials)
- - Dependencies versions update
+ - **BREAKING**: Removed the `integrateOn` method on `Nonlinear`. Now numerical integration algorithms live on their own in the `src/integral` folder
  - Created the `Interpolation` type to work with points interpolation
+ - Moved `NumericalIntegration` into a 'top-level' directoy inside `src/`
  - Added the `LinearInterpolation`, `PolynomialInterpolation`, and `NewtonInterpolation` types
  - Added eigenvalues computation on `Matrix<T>` with the `eigenValues()` method
  - Added inverse matrix computation on `Matrix<T>` with the `inverse()` method
@@ -14,6 +15,7 @@
  - Added the `isIdentity()` method on `Matrix<T>`
  - Added the `trace()` method on `Matrix<T>`
  - Added the `rank()` method on `Matrix<T>`
+ - Dependencies versions update
 
 ## [2.1.3]
